pyproject.toml
[project]
name = "cmf"
version = "0.1.0"
description = ""
authors = [
    {name = "Даниил Смирнов",email = "daniil.horror@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pathlib (>=1.0.1,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "datetime (>=5.5,<6.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pandas-ta (>=0.4.71b0,<0.5.0)",
    "requests (>=2.32.5,<3.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
